重写下载器

+++ MinecraftLaunch/Classes/Models/Event/DownloadProgressChangedEventArgs.cs
@@ -10,12 +10,12 @@ namespace MinecraftLaunch.Classes.Models.Event {
 
         public int TotalCount { get; set; }
 
-        public int TotalBytes { get; set; }
+        public double TotalBytes { get; set; }
 
         public int FailedCount { get; set; }
 
         public int CompletedCount { get; set; }
 
-        public int DownloadedBytes { get; set; }
+        public double DownloadedBytes { get; set; }
     }
 }
